Add a range to the analyzer version
Allow a version range for the analyzer package so that downstream
packages can take advantage of the newer build_resolver package. Of
note, the newer build_resolver resolves dart:ui types.

This also changes the supertype check for mixin-style stores to use the
TypeSystem class. The previous method is deprecated in the latest
analyzer.

Updating the analyzer has the side-effect of not outputting prefixed
types (like ui.Color). It appears that prefixed types weren't working
with non-dart:ui types either, so it is at least consistent now, but
this makes the change breaking.

I did a bit of poking around to figure out how we could support prefixed
types, and there are a few options. For example, we could always rely on
the `_findElementNode` to supply the type name, or we could interrogate
the Element's `linkedNode`.