Upgrade the analyzer to support a range of versions (#309)
* Add a range to the analyzer version

Allow a version range for the analyzer package so that downstream
packages can take advantage of the newer build_resolver package. Of
note, the newer build_resolver resolves dart:ui types.

This also changes the supertype check for mixin-style stores to use the
TypeSystem class. The previous method is deprecated in the latest
analyzer.

Updating the analyzer has the side-effect of not outputting prefixed
types (like ui.Color). It appears that prefixed types weren't working
with non-dart:ui types either, so it is at least consistent now, but
this makes the change breaking.

I did a bit of poking around to figure out how we could support prefixed
types, and there are a few options. For example, we could always rely on
the `_findElementNode` to supply the type name, or we could interrogate
the Element's `linkedNode`.

* Add support for import prefixes across the board

We will now use the file's AST to determine type names in all
situations, rather than asking the Element (which does not include that
information).

This change required a small change in the element visitor — we now
exclude synthetic constructors from consideration, since they don't
appear in the AST. This won't cause problems.

Fixes #130

* Bump package version to 0.3.10