Fix @observable generic types when using import aliases
There were a number of bugs with the previous implementation of the
`LibraryScopedNameFinder`. This resolves them, as well as ensures that a
single code path is followed whether or not the analyzed source code
contains named imports, reducing the potential for future bugs.

The following bugs have been corrected when using named imports:
* Missing type arguments on classes
* Missing type arguments on function typedefs
* Missing prefixes from imported typedefs
* Missing prefixes from implicit type argument bounds

Fixes #367

MR: import_alias_generics