Ensure ObservableStream cleans-up/closes correctly
Previously, observable stream listens to the original when subscribed and pauses when there are no subscriptions and observations.

However, this means it does not receive onDone events from the parent stream, so it cannot properly close its internal subscription

Commit fixes this by enforcing a single subscription only for non-broadcast streams, and sets up a separate internal subscription for broadcast streams, all to ensure we catch the onDone event.

Limitation is that non-broadcast streams that are observed but then unobserved will still not close correctly, but this bug existed since before these changes.