perf: avoid creating strings on disabled spy
We should not create Strings which we're not going to use. In the
counter example we never get to the `conditionallyRunInAction`'s else,
therefore never using the memory and putting more pressure on the GC.

This commit reduces the usage of memory by not creating `actionName`
if we have spy disabled.