Site updates (#981)
* Update documentation and enhance UI

Reformatted and enhanced the presentation of inline code and links in the documentation. Updated versions of dependencies used for building the documentation website. Improved readability of documentation content by enforcing word-wrapping. Made cosmetic changes to the shields/badges on the documentation website for better consistency and readability.

* Update documentation and enhance UI

Reformatted and enhanced the presentation of inline code and links in the documentation. Updated versions of dependencies used for building the documentation website. Improved readability of documentation content by enforcing word-wrapping. Made cosmetic changes to the shields/badges on the documentation website for better consistency and readability.

* more formatting changes

* more formatting changes

* more formatting changes

* responsive updates to the homepage

* changing master -> main in the examples docs

* fixing the skipped css classes