Update package versions and SDK requirements
This commit updates the SDK requirements and versions for various packages in the mobx project. It also corrects a test case in `atom_extensions_test.dart` to properly change state. `mobx_lint_flutter_test` is removed from the build matrix due to failure, and additional package filters are also added for version setting in `melos.yaml`.