Refactor several classes for code readability and upgrade package versions
This commit refactors multiple classes in the mobx, flutter_mobx, and mobx_example libraries to simplify and improve parameter passing. It also upgrades the versions of mobx and flutter_mobx packages. JSON serialization behaviour in the Todo List example of mobx_example is updated to exclude certain attributes from toJSON and fromJSON methods.